Add script for integrating locale into an existing cdclient.sqlite DB (#15)
Originally I was going to write a script that does both fdb&locale -> sqlite conversion, but that was more cumbersome than expected, and also wasn't very useful for existing cdclient.sqlite DBs.

This script is likely still not fully efficient, taking more than a minute to run on my machine whereas the fdb->sqlite script takes a second. Maybe this can be optimized in the future. For now I'm submitting this as is to have *some* implementation out so LU-X SQL API changes are unblocked.